[android] Support safe fastpath for bundledAssets in Expo Client
